🚧 in progress diary update, delete
