+++ src/main/java/com/grepp/diary/app/controller/web/diary/DiaryController.java
@@ -48,7 +48,7 @@ public class DiaryController {
                                                         .collect(Collectors.groupingBy(k -> k.getType().name()));
         model.addAttribute("keywordGroups", grouped);
 
-        return "diary/diary";
+        return "wirte-diary";
     }
 
     @PostMapping
