+++ src/main/java/com/grepp/diary/app/model/diary/repository/DiaryKeywordRepository.java
@@ -1,7 +1,10 @@
 package com.grepp.diary.app.model.diary.repository;
 
+import com.grepp.diary.app.model.diary.entity.Diary;
 import com.grepp.diary.app.model.keyword.entity.DiaryKeyword;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface DiaryKeywordRepository extends JpaRepository<DiaryKeyword, Long> {
+
+    void deleteByDiary(Diary diary);
 }
